Delete user reducer and add join attribute to mission
