+++ src/components/MissionsList/MissionsList.jsx
@@ -19,7 +19,7 @@ const MissionsList = () => {
           id={mission.mission_id}
           name={mission.mission_name}
           description={mission.description}
-          booked={mission.booked}
+          join={mission.join}
         />
       ))}
     </ul>
